implement memory locally

+++ flock_asm/src/compiler.rs
@@ -95,16 +95,16 @@ fn compile_action<'s>(
         Statement::Command1("JOIN", Argument::LiteralNumber(n)) => OpCode::Join(*n).into(),
         Statement::Command0("HALT") => OpCode::Halt.into(),
         Statement::Command1("STORE", arg) => {
-            thunk(move |table| Ok(OpCode::Store(resolve(arg, table)?)))
+            thunk(move |table| Ok(OpCode::Store(resolve(arg, table)? as u64)))
         }
         Statement::Command1("STORE_REL", arg) => {
-            thunk(move |table| Ok(OpCode::StoreRelative(resolve(arg, table)?)))
+            thunk(move |table| Ok(OpCode::StoreRelative(resolve(arg, table)? as u64)))
         }
         Statement::Command1("LOAD", arg) => {
-            thunk(move |table| Ok(OpCode::Load(resolve(arg, table)?)))
+            thunk(move |table| Ok(OpCode::Load(resolve(arg, table)? as u64)))
         }
         Statement::Command1("LOAD_REL", arg) => {
-            thunk(move |table| Ok(OpCode::LoadRelative(resolve(arg, table)?)))
+            thunk(move |table| Ok(OpCode::LoadRelative(resolve(arg, table)? as u64)))
         }
         Statement::Command0("PANIC") => OpCode::Panic.into(),
         s => Err(CompilationError::UnrecognizedStatement(format!("{:?}", s)))?,
